+++ src/main/resources/templates/front/InquiryReview/InquiryWrite.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html xmlns:th="http://www.thymeleaf.org"
       xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
-      layout:decorate="~{admin/layouts/main}">
+      layout:decorate="~{front/layouts/main}">
 <section layout:fragment="content">
     <div>
         <form  method="POST" name="write" th:action="@{/InquiryWrite}"><!--문의 작성 후 넘어갈 html 필요-->
